// src/api/generated/model/portMappingProtocol.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * Cosy API
 * Management API for Cosy (Cost Optimised Server Yard).
 * OpenAPI spec version: v1.0
 */

export type PortMappingProtocol = typeof PortMappingProtocol[keyof typeof PortMappingProtocol];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PortMappingProtocol = {
  TCP: 'TCP',
  UDP: 'UDP',
} as const;